% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basHAL.R
\docType{class}
\name{basHAL}
\alias{basHAL}
\title{Basis adaptive sampling for highly adaptive lasso}
\format{
\code{\link{R6Class}} object.
}
\value{
A basHAL object with methods to run the algorithm
and extract final basis set.
}
\description{
This class implements the basis adaptive sampling algorithm
for highly adaptive lasso.
}
\section{Parameters}{

\itemize{
\item \code{X}: Covariate matrix. Rows are observations and columns are
covariates.
\item \code{y}: Outcome vector. Must be of length equal to the number of
rows in \code{X}.
\item \code{len_candidate_basis_set}: Number of basis functions \eqn{k}
in each candidate basis set.
\item \code{len_final_basis_set}: Number of basis functions \eqn{K} in
the final basis set.
\item \code{max_rows}: Maximum number of rows to sample when fitting
each small model. We recommend setting this to a number smaller than
the number of rows in \code{X} in large data sets for fast computation.
If \code{NULL}, all rows will be used.
\item \code{max_degree}: Maximum degree of interaction allowed in the basis
functions. By default, \code{max_degree=5}. When the number of covariates
\eqn{d} is smaller than 5, the \code{max_degree} is set to \eqn{d}.
\item \code{batch_size}: Number of candidate basis sets to generate
in each iteration.
\item \code{n_batch}: Total number of iterations to run.
\item \code{p}: Proportion of candidate basis sets in each batch that
will be sampled from the updated (posterior) sampling distribution.
The rest will be sampled from the prior distribution.
\item \code{V_folds}: Number of folds to split the data to in cross-validations
when evaluating candidate basis sets.
\item \code{n_cores}: Number of cpu cores to parallel across.
\item \code{seed}: Seed.
\item \code{weight_function}: Weight function, "glmnet" or "glm".
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-basHAL-new}{\code{basHAL$new()}}
\item \href{#method-basHAL-generate_basis_set}{\code{basHAL$generate_basis_set()}}
\item \href{#method-basHAL-sample_basis_set}{\code{basHAL$sample_basis_set()}}
\item \href{#method-basHAL-update_weight}{\code{basHAL$update_weight()}}
\item \href{#method-basHAL-fit_small_glmnet}{\code{basHAL$fit_small_glmnet()}}
\item \href{#method-basHAL-fit_small_glm}{\code{basHAL$fit_small_glm()}}
\item \href{#method-basHAL-evaluate_candidate}{\code{basHAL$evaluate_candidate()}}
\item \href{#method-basHAL-get_top_K}{\code{basHAL$get_top_K()}}
\item \href{#method-basHAL-fit_final_basis_set}{\code{basHAL$fit_final_basis_set()}}
\item \href{#method-basHAL-get_os_loss}{\code{basHAL$get_os_loss()}}
\item \href{#method-basHAL-get_cv_loss}{\code{basHAL$get_cv_loss()}}
\item \href{#method-basHAL-run}{\code{basHAL$run()}}
\item \href{#method-basHAL-clone}{\code{basHAL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-new"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$new(
  X,
  y,
  len_candidate_basis_set,
  len_final_basis_set,
  max_rows,
  max_degree,
  batch_size = 50,
  n_batch = 50,
  p = 0.5,
  V_folds = 5,
  n_cores = 1,
  seed = NULL,
  weight_function = "glmnet",
  family = "gaussian",
  top_K_losses = NULL,
  best_loss = Inf,
  best_basis_set = NULL,
  best_loss_batch = NULL,
  best_loss_traj = NULL,
  loss_prop = 0.5,
  cv_loss = FALSE,
  X_train = NULL,
  X_valid = NULL,
  y_train = NULL,
  y_valid = NULL,
  top_k = FALSE,
  small_fit = "glmnet",
  avg_losses = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-generate_basis_set"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-generate_basis_set}{}}}
\subsection{Method \code{generate_basis_set()}}{
Randomly draw without replacement to form a candidate basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$generate_basis_set()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of \code{Basis} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-sample_basis_set"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-sample_basis_set}{}}}
\subsection{Method \code{sample_basis_set()}}{
Draw without replacement from the updated sampling distribution of
basis functions to form a candidate basis set.
from the current basis sampling distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$sample_basis_set()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of \code{Basis} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-update_weight"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-update_weight}{}}}
\subsection{Method \code{update_weight()}}{
Update the weights of the given basis functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$update_weight(bases, weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bases}}{A list of \code{Basis} objects to update weights for.}

\item{\code{weights}}{A vector of weights to update the basis functions with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-fit_small_glmnet"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-fit_small_glmnet}{}}}
\subsection{Method \code{fit_small_glmnet()}}{
Fit a cross-validated lasso on the training set,
predict using the validation set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$fit_small_glmnet(X_train, y_train, X_valid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_train}}{Training set design matrix.}

\item{\code{y_train}}{Training set response vector.}

\item{\code{X_valid}}{Validation set design matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of predictions, coefficients,
and indices of bases with non-zero coefficients.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-fit_small_glm"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-fit_small_glm}{}}}
\subsection{Method \code{fit_small_glm()}}{
Fit a glm on the training set,
predict using the validation set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$fit_small_glm(basis_set, X_train, y_train, X_valid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-evaluate_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-evaluate_candidate}{}}}
\subsection{Method \code{evaluate_candidate()}}{
Evaluate the performance of the given candidate basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$evaluate_candidate(basis_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis_set}}{A list of \code{Basis} objects to evaluate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of bases indices, calculated weights, and loss.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-get_top_K"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-get_top_K}{}}}
\subsection{Method \code{get_top_K()}}{
Select the K bases with the highest sampling probability
to form the final basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$get_top_K(n_sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_sample}}{The number of bases to select.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of \code{Basis} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-fit_final_basis_set"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-fit_final_basis_set}{}}}
\subsection{Method \code{fit_final_basis_set()}}{
Fit a cross-validated lasso on the final basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$fit_final_basis_set(final_basis_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{final_basis_set}}{The final basis set selected.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A fitted \code{glmnet} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-get_os_loss"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-get_os_loss}{}}}
\subsection{Method \code{get_os_loss()}}{
Get the out-of-sample loss of a given candidate basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$get_os_loss(basis_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis_set}}{A \code{Basis} object for the candidate basis set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The out-of-sample loss of the candidate basis set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-get_cv_loss"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-get_cv_loss}{}}}
\subsection{Method \code{get_cv_loss()}}{
Get the cross-validated loss of a given candidate basis set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$get_cv_loss(basis_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{basis_set}}{A \code{Basis} object for the candidate basis set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cross-validated loss of the candidate basis set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-run"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-run}{}}}
\subsection{Method \code{run()}}{
Run the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$run(verbose = FALSE, plot = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Whether to print progress.}

\item{\code{plot}}{Whether to plot the rate at which the number of bases
explored increases. (FOR TESTING PURPOSES ONLY AT THE MOMENT,
LATER INCORPORATE INTO PROGRESS BAR)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basHAL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-basHAL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basHAL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
