% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis.R
\docType{class}
\name{Basis}
\alias{Basis}
\title{Basis class}
\format{
\code{\link{R6Class}} object.
}
\description{
Class to parameterize a basis function with methods to hash and compare.
}
\section{Parameters}{

\itemize{
\item \code{col_indices}: Integer vector of column indices.
\item \code{knot_points}: Numeric vector of knot points.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Basis-new}{\code{Basis$new()}}
\item \href{#method-Basis-hash}{\code{Basis$hash()}}
\item \href{#method-Basis-equal}{\code{Basis$equal()}}
\item \href{#method-Basis-clone}{\code{Basis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-hash"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-hash}{}}}
\subsection{Method \code{hash()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$hash()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-equal"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-equal}{}}}
\subsection{Method \code{equal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$equal(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Basis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Basis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Basis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
