% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_matrix.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Make design matrix from basis set}
\usage{
make_design_matrix(basis_set, X)
}
\arguments{
\item{basis_set}{A list \code{Basis} objects.}

\item{X}{A vector of values to evaluate the basis functions at.}
}
\value{
A design matrix.
}
\description{
Make design matrix from basis set
}
